requirements.txt
flask
pandas
numpy
PyPDF2
matplotlib
plotly
openai
python-dotenv
werkzeug